backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
dj-database-url>=2.1.0
django-filter>=23.5
psycopg2-binary>=2.9.9
python-decouple>=3.8
django-cors-headers>=4.3.1
drf-spectacular>=0.27.0

# Real-time Pipeline (Feature 002)
channels==4.0.0
channels-redis==4.1.0
paho-mqtt>=2.0.0
redis==5.0.1

# Analytics and Reporting (Feature 003)
reportlab==4.4.9
matplotlib==3.10.8
pillow==12.1.0
numpy==2.4.1

# ML/DL Data Preparation (Feature 004)
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning Models Training (Feature 006)
tensorflow>=2.13.0
